+++ apps/indexer-proxy/proxy/src/metadata/mod.rs
@@ -1,9 +1,9 @@
 mod rpc_evm;
 mod rpc_substrate;
 mod subquery;
-mod thegraph;
+mod subgraph;
 
 pub use rpc_evm::metadata as rpc_evm_metadata;
 pub use rpc_substrate::metadata as rpc_substrate_metadata;
 pub use subquery::metadata as subquery_metadata;
-pub use thegraph::metadata as thegraph_metadata;
+pub use subgraph::metadata as subgraph_metadata;
